test: 테스트 Fixture 적용 및 Service 단위 테스트와 동시성 테스트 분리
